Multiple fixes
I have downloaded your Analysing Digital Images software, and it is
exactly what I need for a measurement task that I have to do – many
thanks.  However, I think that I have come across a flaw in the Spatial
Analysis routine which I thought you would like to know about.

If I go to the 'Select Method to Calibrate the Pixel Size' screen, and
then to 'Scale Present in Image', and draw either a horizontal or
vertical line as my scale then everything works fine.  However, if I
draw a diagonal line for my scale, and then check the length of that
line using the 'line tool' on the next screen, the software does not
give the length correctly.  If, for example, I choose a line at exactly
45 degrees for the scale (i.e., same number of pixels horizontally and
vertically) and specify its length as 1, then the 'line tool' gives the
length of this line as 1.41.  This sounds suspiciously like the square
root of 2.  It seems that the software is just using either the x
coordinate or the y coordinate to determine the number of pixels in the
scale line, although somehow it seems to use whichever is the greater.

- if zoomed in and panned on one window, all windows should have these
settings (spatial - color enhance - mask colors) - and also when
selecting original or enhanced or masked images in each window (lose
zoom and pan settings)

- Enhance colors (predefined enhancements) did not work after selecting
"red only" - until I selected "Original Image" again, then all worked.
Even when I opened another image, I had a similar problem.

- Image masks when zooming in - should wait until click mask button

- When enhance or mask an image, it should open in all windows - bottom
line, if working on an image, it should appear in any window selected
with same zoom and pan settings.

- Similarly, if a measurement tool is selected and used in the spatial
window, it should be there when returning to the window - very time
consuming to recreate certain tools - actually all since the position
selected most likely was there for a very specific reason

- In Old Growth Forest - once spatial tool drawn, should be able to
resize in any of the maps, not just the map drawn in for the original
drawing

-Open recent does not work. (It now does, be sure to check a panel to
see the newly opened picture)