February bug fixes
=======
ADI
=======

Fixed bugs:

-[1h] Diagonal Pixel size calibration: No longer uses either the x or y
coordinate to calibrate length and now correctly calculates the length
of the diagonal selected.
-[4h] Zoom and pan are now preserved through panel changes, including in
SpatialAnalysis when switching between original, enhanced, and masked
image. Same with tools when applicable.
-[1h]Enhance Colors: predefined enhancements didn't work after selecting
the first one. This was one of the original bugs and has been fixed with
the merging of versions
-[0.5h]Mask: Image no longer masks on zoom as requested.
-[1h]In Old Growth Forest, tool can now be manipulated from every image.
-[3h]Open recent didn't work: it now does, be sure to check a panel to
see the newly opened picture. An important number of bugs related to
image format were also fixed.
-[1h]Predefined enhancements now work on the original image, not on the
displayed version unless the "use image as original" button is clicked;
note that pressing that button resets zoom.
-[2h]Enhance Colors had a lot of zoom related bugs and problems that
significantly hindered performance but more importantly slightly
misadjusted zoom and pan values making precision nearly impossible.
Things are rock steady now.
-[3h]In Time Series window, the tool drawn remains when changing
enhancement and values are recalculated automatically. This was a tricky
one!
-[4h]In Spatial analysis, the display for masked pixels was not
implemented at all and setup to simply display the total number of
pixels. I found a nifty little trick to get an accurate number of masked
pixels for each of the tools. This is what caused the bug with the
Polygon tool reported by Francis Precht